requirements.txt
translate
certifi==2018.1.18
chardet==3.0.4
futures==3.1.1
idna==2.6
nltk==3.2.5
numpy==1.14.2
pandas==0.22.0
python-dateutil==2.7.2
pytz==2018.3
requests==2.18.4
six==1.11.0
textblob==0.15.1
urllib3==1.22
